fix(generator): widen int, uint and uint32 to format int64 (#105)
setBasicTypeAndFormat mapped int, uint and uint32 to format int32. int and
uint are 64-bit on every 64-bit target, and uint32 cannot fit in int32 at
all, so generated clients were typed too narrow and intFitsFormat silently
dropped valid examples such as example:"3000000000" on those types. Named
scalars over these builtins inherited the same mapping through
setNamedScalarType, as do their slice items and parameters.

Now int, uint and uint32 emit format int64; uint and uint32 keep the
unsigned minimum: 0. int8, int16, int32, rune, uint8, uint16 and byte stay
int32. uint64 stays int64 with minimum: 0: dropping its format would demote
openapi-generator clients from Long to Integer to fix a mismatch that only
appears at values of 2^63 and above. Leaving int/uint format-less was
rejected for the same reason: openapi-generator maps a format-less integer
to int32. Only format assignments changed; the frozen uint minimum: 0
pre-stamp seam (#54) is untouched, and explicit validate bounds still
overwrite it. The int64 case now compares against goTypeInt64 instead of
the formatInt64 constant, which matched only because both are "int64".

Impact: generated client types for int, uint and uint32 widen from int32 to
int64, and examples above the int32 range on those types are now kept.
kin-openapi enforces a range only for int32, so document validity is
unaffected.

Goldens moved (int32 -> int64 on the affected fields): collision
(quantity), example_tags (limit, bad, count, huge; huge now keeps
example: 3000000000), json_excluded_request (widgets), literal_forms
(limit), path_param and path_param_synth (page). well_known does not move.
example_tags gains an int32 field, Narrow, with example:"3000000000" so the
int32-overflow drop stays golden-locked.

README Known limitations now records that int/uint are documented as
64-bit (wrong for 32-bit builds) and that uint64 values of 2^63 and above
exceed the declared int64 format.

Closes #95

+++ internal/spectest/spectest_test.go
@@ -150,7 +150,7 @@ func TestNamedScalarBuildTaggedWidths(t *testing.T) {
 			amd64:      "type Word int64\ntype UWord uint32\n",
 			i386:       "type Word int64\ntype UWord uint32\n",
 			size:       map[string]any{"type": "integer", "format": "int64"},
-			count:      map[string]any{"type": "integer", "format": "int32", "minimum": 0},
+			count:      map[string]any{"type": "integer", "format": "int64", "minimum": 0},
 			sizesItems: map[string]any{"type": "integer", "format": "int64"},
 		},
 	}
